Remove cancellationToken.js
