--- src/cancellationToken/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "extends": "../tsconfig-base",
-    "compilerOptions": {
-        "types": [
-            "node"
-        ]
-    },
-    "include": ["**/*"]
-}
